qlay_ipc.c : enable writing of MDV files
This allows basic writing to a qlay MDV file. And
writes the result back to disk as the same filename.

There is a hack because for write JS and Minerva use
different hard coded timing contraints for the GAP2
pause. Minervas is shorter so we jump to next state
if a first byte is written.

Signed-off-by: Graeme Gregory <graeme@xora.org.uk>