qlay_io.c : make GAP2 timing on ERASE/WRITE not fixed size
Minerva and JS use slightly different timings for the GAP2 size
on a write. But we can work it out as ERASE comes on at the
beginning of GAP2. Then WRITE comes on at the end.

This means GAP_COUNT has been reduced to 10 to speed up reads
a bit.

Signed-off-by: Graeme Gregory <graeme@xora.org.uk>